+++ src/wai/annotations/isp/polygon_discarder/specifier/__init__.py
@@ -1,0 +1,1 @@
+from ._PolygonDiscarderISPSpecifier import PolygonDiscarderISPSpecifier